feat(windows): continue rework of profile installation

+++ windows/src/support/unload_keyboards/UfrmUnloadKeyboards.pas
@@ -882,7 +882,7 @@ end;
 
 procedure TForm1.cmdITfInputProcessorProfileMgrRegisterProfileClick(Sender: TObject);
 var
-  langid: UINT;
+  langid: WORD;
   guid, clsid: TGUID;
   pInputProcessorProfiles: ITfInputProcessorProfiles;
   pInputProcessorProfileMgr: ItfInputProcessorProfileMgr;
